Adding Maximum Likelihood Objective energy minimizer using BVZ alpha expansion

+++ app/src/main/java/com/lee/kyuhae/john/compphoto/algorithm/maxflow/Arc.java
@@ -14,5 +14,5 @@ public class Arc {
     private Arc next;
     private Arc sister;
 
-    private float residualCapacity;
+    private double residualCapacity;
 }
