+++ app/src/main/java/com/lee/kyuhae/john/compphoto/algorithm/maxflow/Node.java
@@ -17,5 +17,5 @@ public class Node {
     private int distance;
     private boolean sink;
 
-    private float residualCapacity;
+    private double residualCapacity;
 }
